services/game-detector/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.23.2
opencv-python==4.8.0.74
pytesseract==0.3.10
Pillow==10.0.1
numpy==1.26.1
requests==2.31.0
